Extract and Create TaskList class and command class
