+++ src/main/java/Duke/Commands/ExitCommand.java
@@ -1,0 +1,10 @@
+package Duke.Commands;
+
+import Duke.UI.UserInterface;
+
+public class ExitCommand {
+
+    public static void dukeExit() {
+        UserInterface.printDukeExit();
+    }
+}
