--- text-ui-test/duke.txt
@@ -1,3 +1,0 @@
-D | 1 | watch lecture | today
-E | 0 | movie | 7
-T | 1 | buy bread |
