+++ text-ui-test/savedData/duke.txt
@@ -1,0 +1,3 @@
+D | 1 | watch lecture | today
+E | 0 | movie | 7
+T | 0 | buy bread |
